simaen-api/requirements.txt
aniso8601==9.0.1
bidict==0.21.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dask==2021.3.0
distro==1.5.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.2.0
Flask-RESTful==0.3.8
Flask-SocketIO==5.0.1
flask-restx==0.4.0
fsspec==2021.4.0
geojson==2.5.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
locket==0.2.1
MarkupSafe==1.1.1
nose==1.3.7
numpy==1.19.5
packaging==20.9
pandas==1.1.5
partd==1.2.0
psutil==5.8.0
psycopg2-binary==2.8.6
pydotplus==2.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-engineio==4.1.0
python-socketio==5.2.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
scikit-build==0.11.1
scipy==0.19.1
six==1.16.0
toolz==0.11.1
urllib3==1.26.4
virtualenv==15.1.0
Werkzeug==1.0.1
